Add image loading/decoding, steam request wrapper

+++ library/src/lib.rs
@@ -1,5 +1,7 @@
 use serde::{Deserialize, Serialize};
 
+
+
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct Customer {
     pub steam_name: String,
@@ -10,7 +12,7 @@ pub struct Customer {
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct Game {
     pub id: u64,
-    pub app_id: u32,
+    pub app_id: u64,
     pub name: String,
     // pub tags: Vec<String>,
 }
